refactor(utils): lift the poll pacing helpers out of api/event
The widening tick and the failed-poll backoff were private to the command
status poll, but the approval-wait loops in cmd/exec and cmd/worksession need
the same pacing against the same server quota. Moving them to utils lets the
three callers share one rule instead of drifting.

MaxConsecutivePollFailures comes along as the bound a wait rides through failed
polls; no caller uses it yet.

PollMaxBackoffTick is exported because the throttle-budget test in api/event
states its own deadline as a constant expression over the cap, so a literal
there would silently drift from the backoff.

Confidence: high
Scope-risk: narrow
Reversibility: clean
