+++ api/event/event_test.go
@@ -19,6 +19,7 @@ import (
 
 	"github.com/alpacax/alpacon-cli/api"
 	"github.com/alpacax/alpacon-cli/client"
+	"github.com/alpacax/alpacon-cli/utils"
 	"github.com/gorilla/websocket"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -478,53 +479,6 @@ func TestPollCommandExecution_ClientTimeout(t *testing.T) {
 	require.ErrorAs(t, err, &clientTimeout, "deadline expiry must surface a client timeout")
 }
 
-func TestNextPollTick(t *testing.T) {
-	base := time.Second
-	tests := []struct {
-		name    string
-		elapsed time.Duration
-		want    time.Duration
-	}{
-		{name: "base tick while the command is young", elapsed: 0, want: base},
-		{name: "still base at the end of the fast window", elapsed: 9 * base, want: base},
-		{name: "widens once the fast window closes", elapsed: 10 * base, want: 5 * base},
-		{name: "holds through the medium window", elapsed: 59 * base, want: 5 * base},
-		{name: "widest once the medium window closes", elapsed: 60 * base, want: 10 * base},
-		{name: "stays widest", elapsed: 3000 * base, want: 10 * base},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			assert.Equal(t, tt.want, nextPollTick(base, tt.elapsed))
-		})
-	}
-}
-
-func TestNextPollBackoff(t *testing.T) {
-	base := time.Second
-	tests := []struct {
-		name       string
-		attempt    int
-		retryAfter time.Duration
-		want       time.Duration
-	}{
-		{name: "first failure keeps the base tick", attempt: 0, want: base},
-		{name: "doubles per consecutive failure", attempt: 1, want: 2 * base},
-		{name: "still doubling", attempt: 3, want: 8 * base},
-		{name: "capped once doubling passes the cap", attempt: 6, want: 60 * base},
-		{name: "stays capped", attempt: 20, want: 60 * base},
-		{name: "server's Retry-After wins over the schedule", attempt: 0, retryAfter: 3 * base, want: 3 * base},
-		{name: "Retry-After wins even when it is shorter", attempt: 5, retryAfter: 3 * base, want: 3 * base},
-		{name: "Retry-After is capped too", attempt: 0, retryAfter: 3000 * base, want: 60 * base},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			assert.Equal(t, tt.want, nextPollBackoff(base, tt.attempt, tt.retryAfter))
-		})
-	}
-}
-
 // throttleServer replies 429 to the first throttled requests, then serves a
 // terminal status. The returned counter holds the requests it has seen.
 func throttleServer(t *testing.T, throttled int, retryAfter string) (*httptest.Server, *atomic.Int32) {
@@ -662,7 +616,7 @@ func TestPollCommandExecution_ThrottleExtensionIsBoundedByCount(t *testing.T) {
 		elapsed += d
 	}
 	// The loop never waits past its deadline, so the waits sum to it exactly.
-	assert.Equal(t, timeout+pollMaxThrottleExtensions*(pollMaxBackoffTick*tick), elapsed,
+	assert.Equal(t, timeout+pollMaxThrottleExtensions*(utils.PollMaxBackoffTick*tick), elapsed,
 		"the deadline must grow by the capped number of grants, not by the whole duration budget")
 }
 
@@ -704,7 +658,7 @@ func TestPollCommandExecution_ProgressRestoresThrottleExtensions(t *testing.T) {
 	const tick = 100 * time.Microsecond
 	// One timeout is worth exactly one round of capped grants, so a second round is
 	// only reachable through the reset.
-	const timeout = pollMaxThrottleExtensions * (pollMaxBackoffTick * tick)
+	const timeout = pollMaxThrottleExtensions * (utils.PollMaxBackoffTick * tick)
 
 	reqCount := &atomic.Int32{}
 	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
